Finalizacion de los Controladores con los CRUD y operaciones basicas. Aún sin testear algunos controladores.

+++ LogicLayer/CustomerLogic.cs
@@ -21,7 +21,7 @@ namespace LogicLayer
         }
         #endregion
 
-        public Customers CSearch(int ID)
+        public Customers CSearch(long ID)
         {
             return (DbContextSingleton.TransporteContext.Customers.Where(c => c.DocRut == ID).FirstOrDefault());
         }
